docs: rust: add a note about rustup override set
This is just FWIW, but O= changes the working directory before the
build starts.

If you do 'rustup override set' in the source tree and O= points a
directory outside, the version override is not effective.

For example,

  $ rustup override set $(scripts/min-tool-version.sh rustc)
  $ make LLVM=1 O=../build-dir

RUSTUP_TOOLCHAIN works irrespective of the working directory. Anyway,
people will eventually find a way they like.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>