HID: uclogic: Allow to filter certain events
Some UGEE v2 devices, like the XP-PEN Deco Pro SW or the XP-PEN Deco
Mini 7, send an event after probe.

This invalid event is interpreted as a pen report and it moves the
mouse cursor to the 0,0 coordinates.

Allow to create a generic list of events to be ignored per device.

Tested-by: Mia Kanashi <chad@redpilled.dev>
Signed-off-by: José Expósito <jose.exposito89@gmail.com>