rust: rename `Ref` to `Arc`
When `Ref` was first introduced, we still had `alloc`'s `Arc` being used
in the codebase, so we chose a different name. Now that `Arc` is gone,
we can reclaim the name, which is desirable because it is consistent
with userspace.

This is a pure refactor with no functional changes intended.

Signed-off-by: Wedson Almeida Filho <wedsonaf@gmail.com>