rust: support the `?` operator in doctests
Now it is possible to have tests that use the `?` operator:

    /// ```
    /// # use kernel::{spawn_work_item, workqueue};
    /// spawn_work_item!(workqueue::system(), || pr_info!("x"))?;
    /// # Ok::<(), Error>(())
    /// ```

The logic is also simplified: instead of extracting the source
code and the originally-crate-level attributes separately,
wrap everything into an extra scope. This allows to reuse
the `rustdoc`-generated `unwrap()` and `Result` type.

This could break if `rustdoc` happens to emit an attribute that
can only be used at the crate-level, like a `feature(...)` one,
but the previous approach could also break in other ways.

There is a workaround for `Result` not being qualified in
the `rustdoc`-generated code -- it picks up our own `Result`,
and with the simplified logic here we cannot easily move our
`use` of the prelude inside the source code.

Suggested-by: Andreas Hindborg <andreas.hindborg@wdc.com>
Signed-off-by: Miguel Ojeda <ojeda@kernel.org>