scripts: port `rustdoc_test_builder.rs` to Rust
The KUnit doctest generation uses `rustdoc --test-builder` to fetch
the doctests, which runs the test builder once per test.

By porting `rustdoc_test_builder.py` to Rust as a hostprog,
the overhead per test is reduced more than an order of magnitude
(e.g. ~50ms to ~1 ms).

Currently, there are about 100 doctests, which means the Python
script is already responsible for a few seconds of build time.
As more doctests continue to be added, this would become more
and more painful.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>