scripts: port `rustdoc_test_gen.py` to Rust
JSON is not being used now, and the other script is already ported,
so it makes sense to port this one too.

With this, `CONFIG_RUST` does not introduce a Python dependency
(for building), and the remaining Python usage is in
`generate_rust_analyzer.py` which is not needed when building.

The performance difference is again more than an order of magnitude,
but in this case it does not matter much since it runs only once
compared to the builder. Nevertheless, at some point it is likely
there will be many more doctests and shaving a tine bit of time
never hurts.

In fact, from a clean build, this happens to be slower, due to
the compilation step of the script itself.

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>