rust: HID: add HID C API mocks
In order to be able to test the HID subsystem abstractions, we need to
avoid calling the C APIs in the KUnit tests.

Create a private module that mocks the `bindings` functions, import it
only when running the tests:

    #[cfg(CONFIG_KUNIT)]
    use crate::hid::bindings_mock as bindings;

And implement mocks for the functions used by the `module_hid_driver!`
macro to test the register and unregister behavior.

Signed-off-by: José Expósito <jose.exposito89@gmail.com>