rust: HID: add a macro to register HID device ID tables
In order to register a device, the HID subsystem needs to know some
information about it such as the device bus, vendor or product.

In C, this information is represented by the `hid_device_id` structure
and it is stored in `hid_driver->id_table` for registration.

Add a similar Rust structure (`hid::DeviceId`) and the macro
`define_hid_id_table!` to generate the values to store in
`hid_driver->id_table`.

Signed-off-by: José Expósito <jose.exposito89@gmail.com>