drm/vkms: Allow to configure multiple CRTCs
Add a list of CRTC configurations to vkms_config and add as many CRTCs
as configured during output initialization.

For backwards compatibility, create a single CRTC in the default
configuration.

Since writeback support, vblank and composer states are now per CRTC,
extract all the fields to the vkms_crtc structure and allow each
vkms_device to have a list of CRTCs.

Signed-off-by: José Expósito <jose.exposito89@gmail.com>