drm/vkms: Allow to configure multiple connectors
Add a list of connectors to vkms_config and helper functions to add and
remove as many connectors as wanted.

Unlike planes, CRTCs and encoders, connectors can be add and removed
once the device is created. To reflect it, add an "enabled" flag in the
configuration and filter disabled connectors in the
vkms_config_get_connectors() function.

For the moment, changing the enabled status of the connector has no
effect after the device is created, but a future patch will add this
capability.

For backwards compatibility, add one enabled connector to the default
configuration.

A future patch will allow to attach connectors and encoders, but for the
moment there are no changes in the way the output is configured.

Signed-off-by: Louis Chauvet <louis.chauvet@bootlin.com>
Signed-off-by: José Expósito <jose.exposito89@gmail.com>